requirements.txt
cachetools==4.2.2
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.1
oauthlib==3.1.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
requests-oauthlib==1.3.0
rsa==4.7.2
uritemplate==3.0.1